pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "google>=3.0.0",
    "google-genai>=1.60.0",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-nomic>=1.0.1",
    "langchain-ollama>=1.0.1",
    "nomic>=3.9.0",
    "ollama>=0.6.1",
    "openai>=2.16.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "sse>=1.2",
]
